requirements.txt
aiogram==3.15.0
gspread==6.1.4
google-auth==2.37.0
python-dotenv==1.0.1
